+++ packages/google/src/RPA/Cloud/Google/__init__.py
@@ -33,148 +33,40 @@ def import_vault():
 class Google(DynamicCore):
     """`Google` is a library for operating with Google API endpoints.
 
+    **Installation**
+
     Usage requires the following steps:
 
     - Create a GCP project
     - Enable approriate APIs
     - Create credentials (OAuth or service account)
-    - Install rpaframework-google package
+    - Install ``rpaframework-google`` package
+
+    Check the latest package version from `PyPI`_.
 
     **Google authentication**
 
     Authentication for Google is set with `service account JSON file` which can be given to the library
     in three different ways or with `OAuth2 token`, which is used for OAuth authentication.
 
-    Methods when using service account:
+    See `library authentication examples`_ for more information.
 
-    - Method 1 as environment variables, ``GOOGLE_APPLICATION_CREDENTIALS`` with path to service account file.
-    - Method 2 as keyword parameter to ``Init Storage`` for example.
-    - Method 3 as Robocorp vault secret. The vault name and secret key name needs to be given in library init
-      or with keyword ``Set Robocorp Vault``. Secret value should contain JSON file contents.
+    **Basic usage examples**
 
-    Method 1. service account using environment variable
+    **Python**
 
-    .. code-block:: robotframework
+    .. code-block:: python
 
-        *** Settings ***
-        Library   RPA.Cloud.Google
+        from RPA.Cloud.Google import Google
 
-        *** Tasks ***
-        Init Google services
-            # NO parameters for Init Vision, expecting to get JSON
-            # with GOOGLE_APPLICATION_CREDENTIALS environment variable
-            Init Vision
+        library = Google()
+        service_account = '/path/to/service_account.json'
 
-    Method 2. service account with keyword parameter
+        library.init_vision(service_account)
+        library.init_text_to_speech(service_account)
 
-    .. code-block:: robotframework
-
-        *** Settings ***
-        Library   RPA.Cloud.Google
-
-        *** Tasks ***
-        Init Google services
-            Init Speech To Text   /path/to/service_account.json
-
-    Method 3. setting Robocorp Vault in the library init
-
-    .. code-block:: robotframework
-
-        *** Settings ***
-        Library   RPA.Cloud.Google
-        ...       vault_name=googlecloud
-        ...       vault_secret_key=servicecreds
-
-        *** Tasks ***
-        Init Google services
-            Init Storage
-
-    Method 3. setting Robocorp Vault with keyword
-
-    .. code-block:: robotframework
-
-        *** Settings ***
-        Library   RPA.Cloud.Google
-
-        *** Tasks ***
-        Init Google services
-            Set Robocorp Vault   vault_name=googlecloud  vault_secret_key=servicecreds
-            Init Storage    use_robocorp_vault=${TRUE}
-
-    Methods when using OAuth token:
-
-    - Method 1 as keyword parameter ``token_file`` to ``Init Storage`` for example.
-    - Method 2 as Robocorp vault secret. The vault name and secret key name needs to be given in library init
-      or with keyword ``Set Robocorp Vault``. Secret value should contain JSON file contents.
-
-    Method 1. The Google Apps Script and Google Drive services are authenticated using this method.
-
-    .. code-block:: robotframework
-
-        *** Settings ***
-        Library   RPA.Cloud.Google
-
-        *** Variables ***
-        @{SCRIPT_SCOPES}     forms   spreadsheets
-
-        *** Tasks ***
-        Init Google OAuth services
-            Init Apps Script    token_file=oauth_token   ${SCRIPT_SCOPES}
-
-    Method 2. setting Robocorp Vault in the library init
-
-    .. code-block:: robotframework
-
-        *** Settings ***
-        Library   RPA.Cloud.Google
-        ...       vault_name=googlecloud
-        ...       vault_secret_key=oauth
-        ...       cloud_auth_type=token
-
-        *** Tasks ***
-        Init Google services
-            Init Storage
-
-
-    **Creating and using OAuth token file**
-
-    The token file can be created using `credentials.json` by running command:
-
-    ``rpa-google-oauth --credentials <filepath> --service drive`` or
-    ``rpa-google-oauth --credentials <filepath> --scopes drive.appdata,drive.file,drive.install``
-
-    This will start web based authentication process, which outputs the token at the end.
-    Token could be stored into ``Robocorp Vault``.
-
-    Example Vault content.
-
-    .. code-block:: json
-
-        "googlecloud": {
-            "oauth-token": "gANfd123321aabeedYsc"
-        }
-
-    Using the Vault.
-
-    .. code-block:: robotframework
-
-        *** Keywords ***
-        Set up Google Drive authentication
-            Set Robocorp Vault   vault_name=googlecloud
-            ...  vault_secret_key=oauth-token
-            ...  cloud_auth_type=token
-            Init Drive
-
-
-    **Installation**
-
-    This library, ``RPA.Cloud.Google`` is available via **rpaframework-google** package.
-
-    Check the latest package version from `PyPI`_.
-
-    .. _PyPI: https://pypi.org/project/rpaframework-google/
-
-    **Examples**
+        response = library.detect_text('imagefile.png', 'result.json')
+        library.synthesize_speech('I want this said aloud', target_file='said.mp3')
 
     **Robot Framework**
 
@@ -199,20 +91,8 @@ class Google(DynamicCore):
                 Log  ${file}
             END
 
-    **Python**
-
-    .. code-block:: python
-
-        from RPA.Cloud.Google import Google
-
-        library = Google
-        service_account = '/path/to/service_account.json'
-
-        library.init_vision(service_account)
-        library.init_text_to_speech(service_account)
-
-        response = library.detect_text('imagefile.png', 'result.json')
-        library.synthesize_speech('I want this said aloud', target_file='said.mp3')
+    .. _PyPI: https://pypi.org/project/rpaframework-google/
+    .. _library authentication examples: https://github.com/robocorp/rpaframework/blob/master/packages/google/docs/authentication.md
     """  # noqa: E501
 
     ROBOT_LIBRARY_SCOPE = "GLOBAL"