+++ packages/google/src/RPA/Cloud/Google/keywords/drive.py
@@ -63,6 +63,7 @@ class DriveKeywords(LibraryContext):
             use_robocorp_vault=use_robocorp_vault,
             token_file=token_file,
         )
+        return self.service
 
     @keyword(tags=["drive"])
     def upload_drive_file(