+++ packages/google/src/RPA/Cloud/Google/keywords/translation.py
@@ -40,6 +40,7 @@ class TranslationKeywords(LibraryContext):
             use_robocorp_vault,
             token_file,
         )
+        return self.service
 
     @keyword(tags=["translation"])
     def translate(