+++ docs/source/releasenotes.rst
@@ -18,7 +18,7 @@ Latest versions
 28.2.0 - 02 Feb 2024
 --------------------
 
-- Library **RPA.Excel.Application** (:pr:`1148`):
+- Library **RPA.Excel.Application** (:pr:`1149`):
 
   - Add new keyword ``Create Pivot Field`` for creating pivot fields for the pivot table.
   - Add new keyword ``Create Pivot Table`` for creating pivot tables in a spreadsheet.
@@ -28,11 +28,11 @@ Latest versions
   - Add new keyword ``Get Range`` for getting a range in a spreadsheet.
   - Add new keyword ``List Tables`` for listing tables in a spreadsheet.
 
-- For all **RPA.*.Application** libraries (:pr:`1148`): Add new utility keyword for
+- For all **RPA.*.Application** libraries (:pr:`1149`): Add new utility keyword for
   Robot Framework syntax ``Set Object Property``
 
 - Security dependency fix given the bump of ``cryptography`` package to version **42.0.2**.
-  (:pr:`1148`)
+  (:pr:`1149`)
 
 17 Jan 2024
 -----------